// git-search-client/src/types/IUser.ts
export interface IUser{
    login: string
    id: number
    node_id: string
    avatar_url: string
    gravatar_id: string
    html_url: string
    type: string
    site_admin: boolean
    name: string
    company: any
    blog: string
    location: any
    email: any
    hireable: any
    bio: any
    twitter_username: any
    public_repos: number
    public_gists: number
    followers: number
    following: number
    created_at: string
    updated_at: string
}